do not sudo the make install to local dir

+++ cameo/parallel.py
@@ -45,7 +45,8 @@ class MultiprocessingView(Singleton):
         return cpu_count()
 
     def shutdown(self):
-        self.pool.terminate()
+        if not self.pool is None:
+            self.pool.terminate()
 
 
 try:
